services/service-b/requirements.txt
fastapi
uvicorn
strawberry-graphql[fastapi]
sqlalchemy
psycopg2-binary
pydantic-settings
redis
